Cleaned Value Pattern
