+++ src/FlaUI.Core/FlaUI.Core.csproj
@@ -119,7 +119,6 @@
     <Compile Include="Patterns\ITransformPattern.cs" />
     <Compile Include="Patterns\ITransformPatternInformation.cs" />
     <Compile Include="Patterns\IValuePattern.cs" />
-    <Compile Include="Patterns\IValuePatternInformation.cs" />
     <Compile Include="Patterns\IWindowPattern.cs" />
     <Compile Include="Patterns\ITransform2PatternInformation.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
