--- src/FlaUI.Core/Patterns/IValuePatternInformation.cs
@@ -1,11 +1,0 @@
-﻿using FlaUI.Core.Patterns.Infrastructure;
-
-namespace FlaUI.Core.Patterns
-{
-    public interface IValuePatternInformation : IPatternInformation
-    {
-        bool IsReadOnly { get; }
-
-        string Value { get; }
-    }
-}
