+++ src/FlaUI.UIA2/Patterns/ValuePattern.cs
@@ -1,6 +1,5 @@
-﻿using System;
-using FlaUI.Core;
-using FlaUI.Core.Identifiers;
+﻿using FlaUI.Core;
+using FlaUI.Core.Elements.Infrastructure;
 using FlaUI.Core.Patterns;
 using FlaUI.Core.Patterns.Infrastructure;
 using UIA = System.Windows.Automation;
@@ -9,22 +8,19 @@ namespace FlaUI.UIA2.Patterns
 {
     public class ValuePattern : PatternBaseWithInformation<UIA.ValuePattern, ValuePatternInformation>, IValuePattern
     {
-        public static readonly PatternId Pattern = PatternId.Register(AutomationType.UIA2, UIA.ValuePattern.Pattern.Id, "Value");
-        public static readonly PropertyId IsReadOnlyProperty = PropertyId.Register(AutomationType.UIA2, UIA.ValuePattern.IsReadOnlyProperty.Id, "IsReadOnly");
-        public static readonly PropertyId ValueProperty = PropertyId.Register(AutomationType.UIA2, UIA.ValuePattern.ValueProperty.Id, "Value");
+        public ValuePattern(AutomationObjectBase automationObject, UIA.ValuePattern nativePattern) : base(automationObject, nativePattern)
+        {
+            Properties = new ValuePatternProperties();
+        }
 
         IValuePatternInformation IPatternWithInformation<IValuePatternInformation>.Cached
         {
-            get { throw new NotImplementedException(); }
+            get { return Cached; }
         }
 
         IValuePatternInformation IPatternWithInformation<IValuePatternInformation>.Current
         {
-            get { throw new NotImplementedException(); }
-        }
-
-        public ValuePattern(AutomationObjectBase automationObject, UIA.ValuePattern nativePattern) : base(automationObject, nativePattern)
-        {
+            get { return Current; }
         }
 
         protected override ValuePatternInformation CreateInformation(bool cached)
@@ -32,14 +28,22 @@ namespace FlaUI.UIA2.Patterns
             return new ValuePatternInformation(AutomationObject, cached);
         }
 
+        public IValuePatternProperties Properties { get; private set; }
+
         public void SetValue(string value)
         {
             NativePattern.SetValue(value);
         }
+    }
 
-        void IValuePattern.SetValue(string value)
+    public class ValuePatternInformation : ElementInformationBase, IValuePatternInformation
+    {
+        public ValuePatternInformation(AutomationObjectBase automationObject, bool cached) : base(automationObject, cached)
         {
-            throw new NotImplementedException();
         }
+
+        public bool IsReadOnly { get { return Get<bool>(ValuePatternIds.IsReadOnlyProperty); } }
+
+        public string Value { get { return Get<string>(ValuePatternIds.ValueProperty); } }
     }
 }
