+++ src/FlaUI.UIA3/FlaUI.UIA3.csproj
@@ -58,7 +58,6 @@
     <Compile Include="Patterns\TransformPatternIds.cs" />
     <Compile Include="Patterns\TransformPatternInformation.cs" />
     <Compile Include="Patterns\ValuePatternIds.cs" />
-    <Compile Include="Patterns\ValuePatternInformation.cs" />
     <Compile Include="Patterns\WindowPattern.cs" />
     <Compile Include="Patterns\WindowPatternIds.cs" />
     <Compile Include="Tools\NativeConditionConverter.cs" />
